Move pkg/util/slice and update refs
