+++ pkg/kubectl/describe/versioned/describe.go
@@ -72,10 +72,10 @@ import (
 	"k8s.io/kubectl/pkg/util/qos"
 	"k8s.io/kubectl/pkg/util/rbac"
 	resourcehelper "k8s.io/kubectl/pkg/util/resource"
+	"k8s.io/kubectl/pkg/util/slice"
 	storageutil "k8s.io/kubectl/pkg/util/storage"
 	"k8s.io/kubernetes/pkg/kubectl/describe"
 	"k8s.io/kubernetes/pkg/kubectl/scheme"
-	"k8s.io/kubernetes/pkg/kubectl/util/slice"
 )
 
 // Each level has 2 spaces for PrefixWriter
