+++ pkg/kubectl/history.go
@@ -35,9 +35,9 @@ import (
 	"k8s.io/client-go/kubernetes"
 	clientappsv1 "k8s.io/client-go/kubernetes/typed/apps/v1"
 	deploymentutil "k8s.io/kubectl/pkg/util/deployment"
+	sliceutil "k8s.io/kubectl/pkg/util/slice"
 	kapps "k8s.io/kubernetes/pkg/kubectl/apps"
 	describe "k8s.io/kubernetes/pkg/kubectl/describe/versioned"
-	sliceutil "k8s.io/kubernetes/pkg/kubectl/util/slice"
 )
 
 const (
